+++ worm/go.mod
@@ -2,4 +2,9 @@ module github.com/brainsonchain/deepworm
 
 go 1.22.3
 
-require github.com/go-chi/chi v1.5.5
+require (
+	github.com/go-chi/chi v1.5.5
+	go.uber.org/zap v1.27.0
+)
+
+require go.uber.org/multierr v1.10.0 // indirect
